+++ djpcms/contrib/monitor/tests/regression/__init__.py
@@ -1,0 +1,1 @@
+from .stdtests import *