+++ tests/regression/http/tests.py
@@ -9,22 +9,27 @@ class HttpTests(test.TestCase):
     
     def setUp(self):
         self.http = http.get_http()
+        self.Request = self.http.Request
+    
+    def environ(self, method = 'POST', input = b'', path = None):
+        return {'PATH_INFO': path or '/',
+                'REQUEST_METHOD': method,
+                'wsgi.input': http.BytesIO(input)}
         
     def testRequest(self):
-        request = self.http.Request({'PATH_INFO': 'bogus', 'REQUEST_METHOD': 'bogus', 'wsgi.input': http.BytesIO(b'')})
-        self.assertEqual(request.GET.keys(), [])
-        self.assertEqual(request.POST.keys(), [])
-        self.assertEqual(request.COOKIES.keys(), [])
-        self.assertEqual(set(request.META.keys()), set(['PATH_INFO', 'REQUEST_METHOD', 'SCRIPT_NAME', 'wsgi.input']))
-        self.assertEqual(request.META['PATH_INFO'], 'bogus')
-        self.assertEqual(request.META['REQUEST_METHOD'], 'bogus')
-        self.assertEqual(request.META['SCRIPT_NAME'], '')
+        request = self.Request(self.environ('bogus',path='bogus'))
+        self.assertFalse(request.GET)
+        self.assertFalse(request.POST)
+        self.assertFalse(request.COOKIES)
+        self.assertTrue(request.environ)
+        self.assertEqual(request.environ['PATH_INFO'], 'bogus')
+        self.assertEqual(request.environ['REQUEST_METHOD'], 'bogus')
 
     def test_parse_cookie(self):
         self.assertEqual(http.parse_cookie('invalid:key=true'), {})
 
     def test_httprequest_location(self):
-        request = HttpRequest()
+        request = self.Request(self.environ())
         self.assertEqual(request.build_absolute_uri(location="https://www.example.com/asdf"),
             'https://www.example.com/asdf')
 
@@ -35,7 +40,7 @@ class HttpTests(test.TestCase):
 
     def test_near_expiration(self):
         "Cookie will expire when an near expiration time is provided"
-        response = HttpResponse()
+        response = self.http.Response(self.environ())
         # There is a timing weakness in this test; The
         # expected result for max-age requires that there be
         # a very slight difference between the evaluated expiration
@@ -44,28 +49,28 @@ class HttpTests(test.TestCase):
         # 1 second larger. To avoid the problem, put in a quick sleep,
         # which guarantees that there will be a time difference.
         expires = datetime.utcnow() + timedelta(seconds=10)
-        time.sleep(0.001)
+        time.sleep(0.1)
         response.set_cookie('datetime', expires=expires)
         datetime_cookie = response.cookies['datetime']
         self.assertEqual(datetime_cookie['max-age'], 10)
 
     def test_far_expiration(self):
         "Cookie will expire when an distant expiration time is provided"
-        response = HttpResponse()
+        response = self.http.Response(self.environ())
         response.set_cookie('datetime', expires=datetime(2028, 1, 1, 4, 5, 6))
         datetime_cookie = response.cookies['datetime']
         self.assertEqual(datetime_cookie['expires'], 'Sat, 01-Jan-2028 04:05:06 GMT')
 
     def test_max_age_expiration(self):
         "Cookie will expire if max_age is provided"
-        response = HttpResponse()
+        response = self.http.Response(self.environ())
         response.set_cookie('max_age', max_age=10)
         max_age_cookie = response.cookies['max_age']
         self.assertEqual(max_age_cookie['max-age'], 10)
-        self.assertEqual(max_age_cookie['expires'], cookie_date(time.time()+10))
+        self.assertEqual(max_age_cookie['expires'], http.cookie_date(time.time()+10))
 
     def test_httponly_cookie(self):
-        response = HttpResponse()
+        response = self.http.Response(self.environ())
         response.set_cookie('example', httponly=True)
         example_cookie = response.cookies['example']
         # A compat cookie may be in use -- check that it has worked
@@ -73,87 +78,8 @@ class HttpTests(test.TestCase):
         self.assertTrue('; httponly' in str(example_cookie))
         self.assertTrue(example_cookie['httponly'])
 
-    def test_limited_stream(self):
-        # Read all of a limited stream
-        stream = LimitedStream(StringIO('test'), 2)
-        self.assertEqual(stream.read(), 'te')
-        # Reading again returns nothing.
-        self.assertEqual(stream.read(), '')
-
-        # Read a number of characters greater than the stream has to offer
-        stream = LimitedStream(StringIO('test'), 2)
-        self.assertEqual(stream.read(5), 'te')
-        # Reading again returns nothing.
-        self.assertEqual(stream.readline(5), '')
-
-        # Read sequentially from a stream
-        stream = LimitedStream(StringIO('12345678'), 8)
-        self.assertEqual(stream.read(5), '12345')
-        self.assertEqual(stream.read(5), '678')
-        # Reading again returns nothing.
-        self.assertEqual(stream.readline(5), '')
-
-        # Read lines from a stream
-        stream = LimitedStream(StringIO('1234\n5678\nabcd\nefgh\nijkl'), 24)
-        # Read a full line, unconditionally
-        self.assertEqual(stream.readline(), '1234\n')
-        # Read a number of characters less than a line
-        self.assertEqual(stream.readline(2), '56')
-        # Read the rest of the partial line
-        self.assertEqual(stream.readline(), '78\n')
-        # Read a full line, with a character limit greater than the line length
-        self.assertEqual(stream.readline(6), 'abcd\n')
-        # Read the next line, deliberately terminated at the line end
-        self.assertEqual(stream.readline(4), 'efgh')
-        # Read the next line... just the line end
-        self.assertEqual(stream.readline(), '\n')
-        # Read everything else.
-        self.assertEqual(stream.readline(), 'ijkl')
-
-        # Regression for #15018
-        # If a stream contains a newline, but the provided length
-        # is less than the number of provided characters, the newline
-        # doesn't reset the available character count
-        stream = LimitedStream(StringIO('1234\nabcdef'), 9)
-        self.assertEqual(stream.readline(10), '1234\n')
-        self.assertEqual(stream.readline(3), 'abc')
-        # Now expire the available characters
-        self.assertEqual(stream.readline(3), 'd')
-        # Reading again returns nothing.
-        self.assertEqual(stream.readline(2), '')
-
-        # Same test, but with read, not readline.
-        stream = LimitedStream(StringIO('1234\nabcdef'), 9)
-        self.assertEqual(stream.read(6), '1234\na')
-        self.assertEqual(stream.read(2), 'bc')
-        self.assertEqual(stream.read(2), 'd')
-        self.assertEqual(stream.read(2), '')
-        self.assertEqual(stream.read(), '')
-
     def test_stream(self):
-        request = http.Request({'REQUEST_METHOD': 'POST', 'wsgi.input': StringIO('name=value')})
-        self.assertEqual(request.read(), 'name=value')
-
-    def test_read_after_value(self):
-        """
-        Reading from request is allowed after accessing request contents as
-        POST or raw_post_data.
-        """
-        request = http.Request({'REQUEST_METHOD': 'POST', 'wsgi.input': StringIO('name=value')})
+        request = self.Request(self.environ(input = b'name=value'))
+        self.assertEqual(request.raw_post_data, b'name=value')
         self.assertEqual(request.POST, {'name': ['value']})
-        self.assertEqual(request.raw_post_data, 'name=value')
-        self.assertEqual(request.read(), 'name=value')
 
-    def test_value_after_read(self):
-        """
-        Construction of POST or raw_post_data is not allowed after reading
-        from request.
-        """
-        request = http.Request({'REQUEST_METHOD': 'POST', 'wsgi.input': StringIO('name=value')})
-        self.assertEqual(request.read(2), 'na')
-        self.assertRaises(Exception, lambda: request.raw_post_data)
-        self.assertEqual(request.POST, {})
-
-    def test_read_by_lines(self):
-        request = http.Request({'REQUEST_METHOD': 'POST', 'wsgi.input': StringIO('name=value')})
-        self.assertEqual(list(request), ['name=value'])